# server/docs/swagger.yaml
basePath: /
definitions:
...
  user.AddUserInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      userId:
        type: string
    required:
    - email
    - name
    - password
    - userId
    type: object
  user.ChangePasswordResponse:
    properties:
      lastLoginAt:
        type: integer
      token:
        type: string
    type: object
  user.ChangePasswordResponseDTO:
    properties:
      correlationId:
        type: string
      data:
        $ref: '#/definitions/user.ChangePasswordResponse'
      message:
        type: string
      status:
        type: string
    type: object
  user.ChangeUserPasswordInput:
    properties:
      email:
        type: string
      newPassword:
        type: string
      password:
        type: string
    required:
    - email
    - newPassword
    - password
    type: object
...
  user.LoginResponse:
    properties:
      changeUserPassword:
        type: boolean
      lastLoginAt:
        type: integer
      token:
        type: string
    type: object
  user.LoginResponseDTO:
    properties:
      correlationId:
        type: string
      data:
        $ref: '#/definitions/user.LoginResponse'
      message:
        type: string
      status:
        type: string
    type: object
  user.LoginUserInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  user.SignupResponseDTO:
    properties:
      correlationId:
        type: string
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
...
  utils.DefaultResponseDTO:
    properties:
      correlationId:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: RespondNow
  description: RespondNow Server APIs.
  termsOfService: https://www.harness.io/legal/subscription-terms
  title: RespondNow Server
  version: "1.0"
paths:
  /auth/changePassword:
    post:
      consumes:
      - application/json
      description: ChangePassword of RespondNow
      operationId: ChangePassword
      parameters:
      - description: ChangePassword of RespondNow
        in: body
        name: '"changePassword"'
        required: true
        schema:
          $ref: '#/definitions/user.ChangeUserPasswordInput'
      - description: correlationId
        in: query
        name: correlationId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.ChangePasswordResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DefaultResponseDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.DefaultResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DefaultResponseDTO'
      summary: ChangePassword of RespondNow
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login to RespondNow
      operationId: Login
      parameters:
      - description: Login to RespondNow
        in: body
        name: '"login"'
        required: true
        schema:
          $ref: '#/definitions/user.LoginUserInput'
      - description: correlationId
        in: query
        name: correlationId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.LoginResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DefaultResponseDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.DefaultResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DefaultResponseDTO'
      summary: Login to RespondNow
      tags:
      - Auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Signup to RespondNow
      operationId: SignUp
      parameters:
      - description: Signup to RespondNow
        in: body
        name: '"signup"'
        required: true
        schema:
          $ref: '#/definitions/user.AddUserInput'
      - description: correlationId
        in: query
        name: correlationId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.SignupResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DefaultResponseDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.DefaultResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DefaultResponseDTO'
      summary: Signup to RespondNow
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"